+++ renderer/js/courses_index2.js
@@ -80,6 +80,8 @@ document.addEventListener('DOMContentLoaded', () => {
 });
 
 function callApiAndRedirect(id) {
+    const spinner = document.getElementById('spinner');
+    spinner.style.display = 'block'; // Show the spinner
     fetch('https://al-maher.net/api/my_script.php', {
         method: 'POST',
         headers: {
@@ -97,26 +99,32 @@ function callApiAndRedirect(id) {
             fetch(url + queryParams)
                 .then(response => response.json())
                 .then(data => {
-                    // Store the data in local storage so it can be accessed in the next page
-                    localStorage.setItem('courseData3', JSON.stringify(data));
-                    // Then redirect to the new page
-                    window.location.href = 'level4.html';
+                    spinner.style.display = 'none'; // Hide the spinner
+
+                    if (data.rows && data.rows.msg === '0') {
+                        Swal.fire({
+                            title: 'حدث خطأ ما',
+                            text: 'يبدوا أنه لا تتوفر بيانات لهذه الدورة أو حدث خطأ ما يرجي التواصل مع الدعم الفني ',
+                            icon: 'error',
+                            confirmButtonText: 'OK'
+                        });
+                    } else {
+
+                        // Store the data in local storage so it can be accessed in the next page
+                        localStorage.setItem('courseData3', JSON.stringify(data));
+                        // Then redirect to the new page
+                        window.location.href = 'level4.html';
+
+                    }
                 })
                 .catch(error => {
-                    console.error('Error:', error);
-                    Swal.fire({
-                        icon: 'error',
-                        title: 'Oops...',
-                        text: 'Something went wrong! Please Call Al-Maher Support',
-                    });
+                    spinner.style.display = 'none'; // Hide the spinner
+                    Swal.fire('حدث خطأ ما  ', ' يرجي المحاولة مجددا أو تواصل مع الدعم الفني ', 'error');
+
                 });
         })
         .catch(error => {
-            console.error('Error:', error);
-            Swal.fire({
-                icon: 'error',
-                title: 'Oops...',
-                text: 'Something went wrong! Please Call Al-Maher Support',
-            });
+            spinner.style.display = 'none'; // Hide the spinner
+            Swal.fire('حدث خطأ ما  ', ' يرجي المحاولة مجددا أو تواصل مع الدعم الفني ', 'error');
         });
 }
